# bruno-openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise RAG Platform API
  version: 1.0.0
  description: Bruno import file for local Enterprise RAG Platform endpoints.
servers:
  - url: http://localhost:8000
    description: Local backend
paths:
...
  /chat/sessions/{session_id}/messages:
    get:
      summary: Get chat session messages
      description: Requires the session router to be registered in the FastAPI app.
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat messages
  /chat:
    post:
      summary: Chat with citations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
            example:
              query: Summarize the document
              top_k: 5
              session_id: SESSION_ID
              document_id: null
              filename: null
      responses:
        "200":
          description: Chat response with citations
  /chat/stream:
    post:
      summary: Stream chat answer
      description: Returns a text/plain stream. Citations are not included in the streamed response body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
            example:
              query: Summarize the document
              top_k: 5
              session_id: SESSION_ID
              document_id: null
              filename: null
      responses:
        "200":
          description: Plain text stream
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
...
    ChatRequest:
      type: object
      required:
        - query
        - session_id
      properties:
        query:
          type: string
        top_k:
          type: integer
          default: 5
        session_id:
          type: string
        document_id:
          type: string
          nullable: true
        filename:
          type: string
          nullable: true